Move the style sheet manipulation back to CodeFileController
